app/requirements.txt
fastapi==0.110.*
uvicorn==0.29.*
redis==5.0.*
httpx==0.27.*
openai==1.30.*
python-dotenv==1.0.*
jinja2==3.1.*
numpy<2.0.0
chromadb==0.6.3
fitz
pymupdf
python-docx
nltk
sentence-transformers
python-multipart

